requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.4
django-ckeditor==6.7.2
django-crispy-forms==2.3
django-js-asset==3.0.1
pillow==11.1.0
sqlparse==0.5.3
